+++ frontend/src/routes/Realm.tsx
@@ -3,7 +3,7 @@ import React, { ReactElement, useState } from "react";
 import { graphql, loadQuery, useMutation } from "react-relay/hooks";
 import type { RealmQuery, RealmQuery$data } from "./__generated__/RealmQuery.graphql";
 import { useTranslation } from "react-i18next";
-import { LuSunrise, LuInfo, LuSettings, LuPlusCircle, LuPenSquare } from "react-icons/lu";
+import { LuInfo, LuSettings, LuPlusCircle, LuPenSquare } from "react-icons/lu";
 import { WithTooltip, screenWidthAtMost } from "@opencast/appkit";
 
 import { environment as relayEnv } from "../relay";
@@ -127,7 +127,6 @@ const query = graphql`
             isUserRealm
             ownerDisplayName
             children { id }
-            blocks { id }
             canCurrentUserModerate
             ancestors { name path ownerDisplayName }
             parent { id }
@@ -175,37 +174,8 @@ const RealmPage: React.FC<Props> = ({ realm }) => {
                 width: 1,
             }}>{siteTitle}</h1>
         )}
-        {realm.blocks.length === 0 && realm.isMainRoot
-            ? <WelcomeMessage />
-            : <Blocks realm={realm} />}
+        <Blocks realm={realm} />
     </>;
-};
-
-const WelcomeMessage: React.FC = () => {
-    const { t } = useTranslation();
-
-    return (
-        <div css={{
-            maxWidth: 500,
-            marginTop: 32,
-            display: "inline-flex",
-            flexDirection: "column",
-            borderRadius: 4,
-            padding: "8px 16px",
-            gap: 16,
-            alignItems: "center",
-            backgroundColor: COLORS.neutral10,
-            border: `2px dashed ${COLORS.happy0}`,
-        }}>
-            <LuSunrise css={{ marginTop: 8, fontSize: 32, minWidth: 32 }} />
-            <div>
-                <h2 css={{ textAlign: "center", fontSize: 20, marginBottom: 16 }}>
-                    {t("welcome.title")}
-                </h2>
-                <p>{t("welcome.body")}</p>
-            </div>
-        </div>
-    );
 };
 
 const UserRealmNote: React.FC<Props> = ({ realm }) => {
