+++ frontend/src/ui/Blocks/index.tsx
@@ -1,24 +1,28 @@
 import React from "react";
 import { graphql, useFragment } from "react-relay/hooks";
 import { match } from "@opencast/appkit";
+import { useTranslation } from "react-i18next";
+import { LuSunrise } from "react-icons/lu";
 
 import { BlocksData$key } from "./__generated__/BlocksData.graphql";
 import { BlocksRealmData$key } from "./__generated__/BlocksRealmData.graphql";
 import { BlocksBlockData$key } from "./__generated__/BlocksBlockData.graphql";
+import { RealmQuery$data } from "../../routes/__generated__/RealmQuery.graphql";
 import { TitleBlock } from "./Title";
 import { TextBlockByQuery } from "./Text";
 import { SeriesBlockFromBlock } from "./Series";
 import { VideoBlock } from "./Video";
 import { PlayerGroupProvider } from "../player/PlayerGroupContext";
 import { PlaylistBlockFromBlock } from "./Playlist";
+import { COLORS } from "../../color";
 
 
 type BlocksProps = {
-    realm: BlocksData$key;
+    realm: BlocksData$key & RealmQuery$data["realm"];
 };
 
-export const Blocks: React.FC<BlocksProps> = ({ realm: realmRef }) => {
-    const realm = useFragment(graphql`
+export const Blocks: React.FC<BlocksProps> = ({ realm }) => {
+    const realmWithBlocks = useFragment(graphql`
         fragment BlocksData on Realm {
             ... BlocksRealmData
             blocks {
@@ -26,8 +30,11 @@ export const Blocks: React.FC<BlocksProps> = ({ realm: realmRef }) => {
                 ... BlocksBlockData
             }
         }
-    `, realmRef);
-    const { blocks } = realm;
+    `, realm);
+
+    if (realmWithBlocks.blocks.length === 0 && realm.isMainRoot) {
+        return <WelcomeMessage />;
+    }
 
     return (
         <div css={{
@@ -36,8 +43,8 @@ export const Blocks: React.FC<BlocksProps> = ({ realm: realmRef }) => {
             rowGap: 32,
         }}>
             <PlayerGroupProvider>
-                {blocks.map(
-                    block => <Block key={block.id} realm={realm} block={block} />,
+                {realmWithBlocks.blocks.map(
+                    block => <Block key={block.id} realm={realmWithBlocks} block={block} />,
                 )}
             </PlayerGroupProvider>
         </div>
@@ -79,4 +86,31 @@ export const Block: React.FC<BlockProps> = ({ block: blockRef, realm }) => {
             "PlaylistBlock": () => <PlaylistBlockFromBlock fragRef={block} basePath={basePath} />,
         })}
     </div>;
+};
+
+const WelcomeMessage: React.FC = () => {
+    const { t } = useTranslation();
+
+    return (
+        <div css={{
+            maxWidth: 500,
+            marginTop: 32,
+            display: "inline-flex",
+            flexDirection: "column",
+            borderRadius: 4,
+            padding: "8px 16px",
+            gap: 16,
+            alignItems: "center",
+            backgroundColor: COLORS.neutral10,
+            border: `2px dashed ${COLORS.happy0}`,
+        }}>
+            <LuSunrise css={{ marginTop: 8, fontSize: 32, minWidth: 32 }} />
+            <div>
+                <h2 css={{ textAlign: "center", fontSize: 20, marginBottom: 16 }}>
+                    {t("welcome.title")}
+                </h2>
+                <p>{t("welcome.body")}</p>
+            </div>
+        </div>
+    );
 };
